♻️ Externalize routes of server
